requirements.txt
requests==2.22.0
Django==2.2.4
djangorestframework==3.10.2
django-cors-headers==2.0.1
Markdown==3.1.1
psycopg2==2.8.3
PyJWT==1.7.1
oss2==2.8.0
shortuuid==0.5.0
django_redis==4.10.0
uWSGI==2.0.18